Add test case for java extensions.
