+++ README.md
@@ -89,6 +89,8 @@ For information on getting started, refer to the [CONTRIBUTING.md instructions](
 
 Continuous Integration builds can be installed from http://download.jboss.org/jbosstools/jdt.ls/staging/. Download the most recent `java-<version>.vsix` file and install it by following the instructions [here](https://code.visualstudio.com/docs/editor/extension-gallery#_install-from-a-vsix).
 
+Also, you can contribute your own VSCode extension to enhance the existing features by following the instructions [here](https://github.com/redhat-developer/vscode-java/wiki/Contribute-a-Java-Extension).
+
 Feedback
 ===============
 * File a bug in [GitHub Issues](https://github.com/redhat-developer/vscode-java/issues).
